Include state in myProfile page & put state in app.js
