+++ src/components/MyProfile.js
@@ -1,7 +1,20 @@
 import React from 'react';
+import { useSelector } from 'react-redux';
 
-const MyProfile = () => (
-  <h1>MyProfile</h1>
-);
+const MyProfile = () => {
+  const { rocketData } = useSelector((state) => state.rockets);
+  // console.log(rocketData);
+  return (
+    <>
+      <p>Rockets</p>
+      <ul>
+        {rocketData.map((rocket) => rocket.reserved && (
+          <li key={rocket.id}>{rocket.name}</li>
+        ))}
+      </ul>
+
+    </>
+  );
+};
 
 export default MyProfile;
